requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
aws-sdk-signers==0.0.3
aws-sdk-bedrock-runtime==0.0.2
awscrt==0.27.4
bidict==0.23.1
reportlab==4.0.7
blinker==1.9.0
boto3==1.40.27
botocore==1.40.27
certifi==2025.7.9
click==8.2.1
colorama==0.4.6
docstring-parser==0.16
Flask==3.1.1
Flask-SocketIO==5.5.1
python-docx==1.1.0
openpyxl==3.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
ijson==3.4.0
importlib-metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
mcp==1.16.0
opentelemetry-api==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
pillow==11.3.0
PyPDF2==3.0.1
PyYAML==6.0.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic-core==2.33.2
python-dateutil==2.9.0
python-dotenv==1.1.1
python-engineio==4.12.2
python-multipart==0.0.20
python-socketio==5.13.0
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
s3transfer==0.14.0
simple-websocket==1.1.0
six==1.17.0
smithy-aws-core==0.0.3
smithy-aws-event-stream==0.0.1
smithy-core==0.0.2
smithy-http==0.0.1
smithy-json==0.0.1
sniffio==1.3.1
sse-starlette==2.4.1
starlette==0.47.1
strands-agents>=1.0.0
strands-agents-tools>=0.2.0
typing-inspection==0.4.1
typing-extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
watchdog==6.0.0
Werkzeug==3.1.3
wsproto==1.2.0
zipp==3.23.0
bedrock-agentcore>=0.1.3
bedrock-agentcore-starter-toolkit>=0.1.10